Refactor into db and handler packages
